[skip] feat: display branch points as a yellow bar line/major grid
Notice that currently yellow major grid will be obcured
by the grey lines of BPM or HS changes.

This does not redo 49981bf35827935c3f5beba21f8fc39ae21865d7 because the feature addition has been nullified.
However, this is still relevant as the yellow bar is expected on branch points and not on condition reset points,
but fixing this requires behavior changes and will be done later.